% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_tests.R
\name{b_hierarchical_test}
\alias{b_hierarchical_test}
\title{Bayesian hierarchical correlated t-test}
\usage{
b_hierarchical_test(df, baseline, algorithm = NULL, measure = NULL,
  rho = 0.1, compare = NULL, std.upper = 1000, d0.lower = NULL,
  d0.upper = NULL, prob = 0.95, alpha.lower = 0.5, alpha.upper = 5,
  beta.lower = 0.05, beta.upper = 0.15, rope = c(-0.01, 0.01),
  nsim = 2000, parallel = TRUE, stan.output.file = NULL,
  nchains = 8, seed = as.numeric(Sys.time()), adapt_delta = 0.8,
  max_treedepth = 10)
}
\arguments{
\item{df}{Input data frame.}

\item{baseline}{First algorithm.}

\item{algorithm}{Second algorithm. If not defined, every algorithm will be 
tested against baseline.}

\item{measure}{Measure column.}

\item{rho}{Correlation factor.}

\item{compare}{Defines whether the baseline should be tested for either 
being better ('better') or being just as good ('equal') as the other 
algorithms. If not defined, the default is to test for 'better'.}

\item{std.upper}{Factor to set the upper bound for both sigma_i and sigma_0.}

\item{d0.lower}{Lower bound for the prior for mu_0. If not provided, 
the smallest observed difference is used.}

\item{d0.upper}{Upper bound for the prior for mu_0. If not provided, 
the biggest observed difference is used.}

\item{prob}{Probability, which the decision that the Baseline is better than 
the algorithm is based on. The default is 0.95.}

\item{alpha.lower}{Lower bound for the (uniform) prior for the alpha 
hyperparameter. Default value set at 0.5.}

\item{alpha.upper}{Upper bound for the (uniform) prior for the alpha 
hyperparameter. Default value set at 0.5.}

\item{beta.lower}{Lower bound for the (uniform) prior for the beta 
hyperparameter. Default value set at 0.5.}

\item{rope}{Interval for the difference considered as 'irrelevant'.}

\item{nsim}{Number of samples (per chain) used to estimate the posterior 
distribution. Note that, by default, half the simulations are used for 
the burn-in.}

\item{parallel}{Logical value. If \code{true}, Stan code is executed in 
parallel.}

\item{stan.output.file}{String containing the base name for the output files 
produced by Stan. If \code{NULL}, no files are stored.}

\item{seed}{Optional parameter used to fix the random seed.}

\item{beta.lower}{Upper bound for the (uniform) prior for the beta 
hyperparameter. Default value set at 0.5.}

\item{z0}{Position of the pseudo-observation associated to the prior 
Dirichlet Process. The default value is set to 0 (inside the rope).}

\item{nchain}{Number of MC chains to be simulated. As half the simulations 
are used for the warm-up, the total number of simulations will 
be \code{nchain}*\code{nsim}/2.}
}
\value{
A list containing the following components:
\itemize{
    \item{code{measure}} A string with the name of the measure column used.
    \item{code{method}} A string with the name of the method used.
    \item{code{baseline}} A string with the name of the baseline algorithm.
    \item{code{posteriror_probabilities}} A data frame with one row for every 
        algorithm that is compared to the baseline. The columns show the 
        posterior probabilities and whether significance appears.
}
}
\description{
This function implements a Bayesian hierarchical test. The performance of 
    one baseline algorithm on multiple data set is compared to either one or
    multiple algorithms.
}
\details{
The test has first been implemented in scmamp. 
    Note that if no measure column is defined per default the first column 
    defined as measure_* in the data frame is used. The default of rho is 
    0.1.
}
\examples{
    results <- b_hierarchical_test(df= test_benchmark_small, 
    baseline = "algo_1", algorithm = "algo_2",  rho=0.1, 
    rope=c(-0.01, 0.01), nsim=2000,  nchains=5)
}
\references{
\url{https://github.com/b0rxa/scmamp}
}
