% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_tests.R
\name{b_signed_rank_test}
\alias{b_signed_rank_test}
\title{Bayesian Signed Rank test}
\usage{
b_signed_rank_test(df, problemset = NULL, baseline, compare = NULL,
  algorithm = NULL, measure = NULL, s = 0.5, z_0 = 0,
  weights = NULL, mc_samples = 1e+05, rope = c(-0.01, 0.01))
}
\arguments{
\item{df}{Input data frame.}

\item{problemset}{Problem set on which the test should be performed.}

\item{baseline}{First algorithm.}

\item{algorithm}{Second algorithm.}

\item{measure}{Measure column.}

\item{s}{Prior pseudo-observation probability.}

\item{z_0}{Prior pseudo-observation.}

\item{weights}{A-priori weights.}

\item{mc_samples}{Number of samples of the distribution.}

\item{rope}{Region of practical equivalence.}
}
\value{
A list containing the following components:
\item{code{measure}}{A string with the name of the measure column used.}
\item{code{method}}{A string with the name of the method used.}
\item{code{posteriror_probabilities}}{A vector with the left, rope and right 
probabilities.}
}
\description{
This function implements the Bayesian version of the signed rank test.
}
\details{
The test has first been implemented in rNPBST. 
For testing over multiple datasets, don't specify the problem set argument 
in the function. 
Note that the default value for measure is the first measure column in the 
data frame.
}
\references{
\url{https://github.com/JacintoCC/rNPBST}
}
