% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_structure.R
\name{check_column_names}
\alias{check_column_names}
\title{Check Columns of data frame}
\usage{
check_column_names(df)
}
\arguments{
\item{df}{(`list`)\cr Data frame containing the performane measure.}
}
\value{
(`logical`)\cr TRUE if test was successful.
}
\description{
Check if data frame only contains legit columns and whether any
    compulsory columns exist.
}
\examples{
    check_column_names(test_benchmark_small)
}
