% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_structure.R
\name{check_structure}
\alias{check_structure}
\title{Check Structure}
\usage{
check_structure(df)
}
\arguments{
\item{df}{(`list`)\cr Data frame containing the performane measure.}
}
\value{
(`logical`)\cr TRUE if test was successful.
}
\description{
Check if the structure of the data frame satisfies the 
    requirements.
}
\examples{
    check_structure(test_benchmark_small)
}
