% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_function.R
\name{data_summary}
\alias{data_summary}
\title{Summary}
\usage{
data_summary(df)
}
\arguments{
\item{df}{input data frame}
}
\value{
A vector containing the columns names and number of rows
}
\description{
Short summary of the data frame, 
  including the columns names and number of rows
}
