% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_transformation}
\alias{data_transformation}
\title{Data transformation}
\usage{
data_transformation(df, algo, measure)
}
\arguments{
\item{df}{(`list`)\cr Data frame containing the performane measure.}

\item{measure}{(`character`)\cr Name of the 'measure' column. If not 
defined, the first 'measure' column in the data frame is used.}

\item{algorithm}{(`character`)\cr Second algorithm. Value in 'algorithm' 
column. If not defined, the baseline is tested against all algorithms 
in the data frame.}
}
\value{
(`matrix`)\cr Matrix.
}
\description{
Extract a data matrix out of the original dataset. The matrix 
    contains all observations for each replication in each problem set. 
    The replications are stored in the rows, while the columns are according  
    to the problem sets.
}
