% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_hoc_tests.R
\name{friedman_post}
\alias{friedman_post}
\title{Friedman's post hoc test}
\usage{
friedman_post(df, measure = NULL, control = NULL)
}
\arguments{
\item{df}{(`list`)\cr Data frame containing the performane measure.}

\item{measure}{(`character`)\cr Name of the 'measure' column. If not 
defined, the first 'measure' column in the data frame is used.}

\item{control}{(`character`)\cr The name of the control algorithm. If this 
parameter is not provided, all algorithms are compared against each 
other.}
}
\value{
A list containing the following components: 
\itemize{
    \item{\code{measure}} (`character`)\cr A string with the name of the 
        measure column. 
    \item{\code{method}} (`character`)\cr A string with the name of the 
        method. 
    \item{code{matrix}} (`matrix`)\cr A matrix with all the pair wise raw p-values. 
}
}
\description{
This function implements a Friedman post hoc test. It computes 
the raw p-values for the post hoc based on Friedman's test.
}
\details{
The test has first been implemented in scmamp.
}
\examples{
    friedman_post(test_benchmark)
}
\references{
\url{https://github.com/b0rxa/scmamp}
}
