% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data_frame_htest}
\alias{get_data_frame_htest}
\title{Get test results (for frequentist tests)}
\usage{
get_data_frame_htest(k, p_value, test, statistic)
}
\arguments{
\item{k}{algorithm tested against the baseline.}

\item{p_value}{the p-value for the test.}

\item{test}{the hypothesis test.}

\item{statistic}{the value of the statistic used in the test.}
}
\value{
List.
}
\description{
This function collects the part of the results shown in the data 
    frame.
}
