% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data_frame_seq}
\alias{get_data_frame_seq}
\title{Get test results (for sequential tests)}
\usage{
get_data_frame_seq(k, posterior, repls)
}
\arguments{
\item{k}{algorithm tested against the baseline.}

\item{posterior}{Call for the posterior probabilities of the bayesian tests.}

\item{repls}{Number of replications used until a decision is made.}
}
\value{
List.
}
\description{
This function collects the part of the results shown in the data 
    frame.
}
