% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_extras_scmamp}
\alias{get_extras_scmamp}
\title{Get extras from tests (for tests from scmamp)}
\usage{
get_extras_scmamp(x, ...)
}
\arguments{
\item{x}{result of test in function.}
}
\value{
List.
}
\description{
This function collects additional information from the test 
    summed up as "extras".
}
