% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_main_columns_count}
\alias{get_main_columns_count}
\title{Count main columns}
\usage{
get_main_columns_count(df)
}
\arguments{
\item{df}{(`list`)\cr Data frame containing the performane measure.}
}
\value{
(`list`)\cr A numeric vector displaying the number of main columns.
}
\description{
Count the number of main columns in the data frame.
}
