% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_measure_columns}
\alias{get_measure_columns}
\title{Measure Columns}
\usage{
get_measure_columns(df)
}
\arguments{
\item{df}{input data frame}
}
\value{
A vector containing all measure columns.
}
\description{
Get the list of names for all measure columns within the data frame.
}
