% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_measure_columns_count}
\alias{get_measure_columns_count}
\title{Count Measure Columns}
\usage{
get_measure_columns_count(df)
}
\arguments{
\item{df}{input data frame}
}
\value{
A numeric vector displaying the number of columns containing measures.
}
\description{
Get the number parameter columns in the data frame.
}
