% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_parameter_columns_count}
\alias{get_parameter_columns_count}
\title{Count Parameter Columns}
\usage{
get_parameter_columns_count(df)
}
\arguments{
\item{df}{input data frame}
}
\value{
A numeric vector displaying the number of parameter columns.
}
\description{
Get the number parameter columns in the data frame.
}
