% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_probabilities}
\alias{get_probabilities}
\title{Get probabilities}
\usage{
get_probabilities(result, compare, prob)
}
\arguments{
\item{result}{(`list`)\cr Data frame with all the results within in the 
tests.}

\item{compare}{(`character`)\cr Defines if one algorithm needs to perform 
better ({\code{better}}) for decisions based on the posterior 
distribution or whether it is sufficient to perform not worse 
({\code{equal}}).}

\item{prob}{(`double`)\cr Threshold probability that decision rely on. 
Default is 0.95.}
}
\value{
(`list`)\cr Column of result data frame.
}
\description{
This function defines the probabilities resulting from the
    Bayesian tests for each algorithm that is tested against the baseline.
}
