% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_replication}
\alias{get_replication}
\title{Build replication}
\usage{
get_replication(i, df)
}
\arguments{
\item{i}{(`double`)\cr Number of replication in data frame.}

\item{df}{(`list`)\cr Data frame containing the performane measure.}
}
\value{
(`list`)\cr Data frame containing defined the number of replication.
}
\description{
When not defined by user, it calls the replication in a 
    complete data frame provided by the user. Can otherwise be used to build 
    replications during testing procedure.
}
