% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_replication_count}
\alias{get_replication_count}
\title{Check number of replication}
\usage{
get_replication_count(x, y)
}
\arguments{
\item{x}{(`list`)\cr First sample.}

\item{y}{(`list`)\cr Second sample.}
}
\value{
(`logical`)\cr TRUE if both samples are of same length.
}
\description{
Check if number of observations in the first and second 
    sample are equal.
}
