% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_replications}
\alias{get_replications}
\title{Build replications}
\usage{
get_replications(i, df)
}
\arguments{
\item{i}{(`double`)\cr Number of replications in data frame.}

\item{df}{(`list`)\cr Data frame containing the performane measure.}
}
\value{
(`list`)\cr Data frame containing defined the number of replications.
}
\description{
When not defined by user, it calls the replications in a 
complete data frame provided by the user. Can otherwise be used to build 
replications during testing procedure.
}
