% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_replications_count}
\alias{get_replications_count}
\title{Check number of replications}
\usage{
get_replications_count(x, y)
}
\arguments{
\item{x}{First sample.}

\item{y}{Second sample.}
}
\value{
TRUE if both samples are of same length.
}
\description{
Check if number of observations in the first and second 
    sample are equal.
}
