% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_results}
\alias{get_results}
\title{Get test results (Bayesian tests)}
\usage{
get_results(baseline, method, measure, data = NULL, extra = NULL)
}
\arguments{
\item{baseline}{Baseline algorithms that every algorithm can be compared to.}

\item{method}{Bayesian test.}

\item{measure}{Measure column.}

\item{data}{Data frame containing the posterior probabilities.}

\item{extra}{Any extra arguments needed to build the plots.}
}
\value{
List.
}
\description{
This function collects the results of the Bayesian tests.
}
