% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_results_htest}
\alias{get_results_htest}
\title{Get test results (Frequentist tests)}
\usage{
get_results_htest(baseline = NULL, method, measure, data = NULL, matrix = NULL)
}
\arguments{
\item{baseline}{(`character`)\cr First algorithm. Value in 'algorithm'
column.}

\item{method}{(`character`)\cr Bayesian test.}

\item{measure}{(`character`)\cr Name of the 'measure' column. If not 
defined, the first 'measure' column in the data frame is used.}

\item{data}{(`list`)\cr Data frame containing the posterior probabilities.}

\item{matrix}{(`matrix`)\cr A matrix with all the pair wise differences of 
average rankings.}
}
\value{
(`list`)\cr List.
}
\description{
This function collects the results of the Frequentist tests.
}
