% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_rows}
\alias{get_rows}
\title{Get relevant rows in data frame (for sequential tests)}
\usage{
get_rows(result)
}
\arguments{
\item{result}{Data frame with all the results within in the tests.}
}
\value{
Data frame with just the relevant rows.
}
\description{
This function drops the non-final results during the sequential 
    testing process.
}
