% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_threshold}
\alias{get_threshold}
\title{Get threshold}
\usage{
get_threshold(posterior, compare)
}
\arguments{
\item{posterior}{(`character`)\cr Call for the posterior probabilities of the 
Bayesian tests.}

\item{compare}{(`character`)\cr Defines if one algorithm needs to perform 
better ({\code{better}}) for decisions based on the posterior 
distribution or whether it is sufficient to perform not worse 
({\code{equal}}).}
}
\value{
(`list`)\cr List.
}
\description{
This function defines the threshold for making any decisions 
    depending on the calculated posterior probabilities.
}
