% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_function.R
\name{na_drop}
\alias{na_drop}
\title{Drop NAs by groups}
\usage{
na_drop(df, check_var = NULL, measure = NULL)
}
\arguments{
\item{df}{(`list`)\cr Data frame containing the performane measure.}

\item{check_var}{(`character`)\cr Column in data frame used to check for NAs. 
Either "problem" (default) or "algorithm".}

\item{measure}{(`character`)\cr Name of the 'measure' column. If not 
defined, the first 'measure' column in the data frame is used.}
}
\value{
(`list`)\cr New data frame without NAs.
}
\description{
Drop group of rows that contain any NA depending on values of check_var.
}
