% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{paste_algo_pars}
\alias{paste_algo_pars}
\title{Paste algorithm and parameter}
\usage{
paste_algo_pars(algorithm, parameter_algorithm)
}
\arguments{
\item{algorithm}{Algorithm in data frame.}

\item{parameter_algorithm}{Algorithm parameter in data frame.}
}
\value{
New algorithm name, combining algorithm and its parameter.
}
\description{
If there is a parameter additional to the algorithm, both can be combined, 
when using this function.
}
