% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{Boxplot}
\usage{
plot_boxplot(df, measure = NULL)
}
\arguments{
\item{df}{(`list`)\cr Data frame containing the performane measure.}

\item{measure}{(`character`)\cr Name of the 'measure' column. If not 
defined, the first 'measure' column in the data frame is used.}
}
\value{
(`list`) \cr List containing a `gg` object.
}
\description{
This function plots the performance of every algorithm in each 
    data frame in boxplots.
}
\details{
The test has first been implemented in scmamp.
}
\examples{
    plot_boxplot(df = test_benchmark_small)
}
\references{
\url{https://github.com/b0rxa/scmamp}
}
