% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_cd}
\alias{plot_cd}
\title{Critical differences plot}
\usage{
plot_cd(df, measure = NULL, alpha = 0.05, cex = 0.75, ...)
}
\arguments{
\item{df}{(`list`)\cr Data frame containing the performane measure.}

\item{measure}{(`character`)\cr Name of the 'measure' column. If not 
defined, the first 'measure' column in the data frame is used.}

\item{alpha}{(`double`)\cr Significance level to get the critical difference.}

\item{cex}{(`double`)\cr Numeric value to control the size of the font.}

\item{...}{(any)\cr Additional arguments.}
}
\value{
(`list`) \cr List containing a `gg` object.
}
\description{
This function implements the critical difference plots 
    introduced in Demsar (2006).
}
\details{
The test has first been implemented in scmamp. By default, the alpha 
    value is 0.05 and the default for cex is 0.75.
}
\examples{
    plot_cd(test_benchmark)
}
\references{
\url{https://github.com/b0rxa/scmamp}
}
