% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_densities}
\alias{plot_densities}
\title{Plot Densities}
\usage{
plot_densities(df, measure = NULL)
}
\arguments{
\item{df}{(`list`)\cr Data frame containing the performane measure.}

\item{measure}{(`character`)\cr Name of the 'measure' column. If not 
defined, the first 'measure' column in the data frame is used.}
}
\value{
(`list`) \cr List containing a `gg` object.
}
\description{
This function estimates and plots the densities of the 
performances of each algorithm in the data frame
}
\details{
The test has first been implemented in scmamp.
}
\examples{
    plot_densities(test_benchmark_small)
}
\references{
\url{https://github.com/b0rxa/scmamp}
}
