% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_triangles}
\alias{plot_triangles}
\title{Plot posterior Dirichlet distribution}
\usage{
plot_triangles(df, method = "Sign", problemset = NULL, learner_a,
  learner_b, measure = NULL, parameter_algorithm = NULL, s = 0.5,
  z_0 = 0, weights = NULL, mc_samples = 1e+05, rope = c(-0.01,
  0.01), points = 10000)
}
\description{
Plot posterior Dirichlet distribution
}
