% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_structure.R
\name{print.b_test}
\alias{print.b_test}
\title{Table test results (Bayesian tests)}
\usage{
\method{print}{b_test}(x, ...)
}
\arguments{
\item{x}{(`list`)\cr Test results.}

\item{...}{(any)\cr Additional arguments.}
}
\description{
Create a list containing the Bayesian test results based on a 
    generic function.
}
