% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_structure.R
\name{print.btest}
\alias{print.btest}
\title{Table test results (Bayesian tests)}
\usage{
\method{print}{btest}(x, ...)
}
\description{
Create a list containing the Bayesian test results based on a 
    generic function.
}
