% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_structure.R
\name{print.h_test}
\alias{print.h_test}
\title{Table test results (Frequentist tests)}
\usage{
\method{print}{h_test}(x, ...)
}
\description{
Create a list containing the Frequentist test results based on a 
    generic function.
}
