% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_structure.R
\name{print.h_test_small}
\alias{print.h_test_small}
\title{Table test results (Frequentist tests)}
\usage{
\method{print}{h_test_small}(x, ...)
}
\arguments{
\item{x}{(`list`)\cr Test results.}

\item{...}{(any)\cr Additional arguments.}
}
\description{
Create a list containing the Frequentist test results based on a 
    generic function. Not containing the Baseline.
}
