% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_structure.R
\name{print.nemenyi}
\alias{print.nemenyi}
\title{Table test results (Nemenyi tests)}
\usage{
\method{print}{nemenyi}(x, ...)
}
\arguments{
\item{x}{(`list`)\cr Test results.}

\item{...}{(any)\cr Additional arguments.}
}
\description{
Create a list containing the Nemenyi test results based on a 
    generic function.
}
