% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_sequential_tests.R
\name{seq_b_signed_rank_test}
\alias{seq_b_signed_rank_test}
\title{Sequential Bayesian Signed Rank test}
\usage{
seq_b_signed_rank_test(problem = NULL, baseline, algorithm = NULL,
  measure = NULL, compare = NULL, s = 0.5, z_0 = 0,
  weights = NULL, mc_samples = 1e+05, rope = c(-0.01, 0.01),
  max_repls = 20, prob = 0.95, min_repls = 5, ...)
}
\arguments{
\item{problem}{Problem set on which the test should be performed.}

\item{baseline}{First algorithm.}

\item{algorithm}{Second algorithm. If not defined, every algorithm will be 
tested against baseline.}

\item{measure}{Measure column.}

\item{compare}{Defines whether the baseline should be tested for either 
being better ('better') or being just as good ('equal') as the other 
algorithms. If not defined, the default is to test for 'better'.}

\item{s}{Prior pseudo-observation probability.}

\item{z_0}{Prior pseudo-observation.}

\item{weights}{A-priori weights.}

\item{mc_samples}{Number of samples of the distribution.}

\item{rope}{Region of practical equivalence.}

\item{max_repls}{Maximum number of replications that should be build in 
get_replications. Or maximum number of replications in data frame, if 
a data frame that has already been built is being used.}

\item{prob}{Probability, which the decision that the Baseline is better than 
   the algorithm is based on. The default is 0.95. 
@param min_repls The minimum number of replications that is either generated 
    or used to perform the first Bayesian test in the sequential approach. 
    To prevent an bias through early stopping a minimum number of 5 runs 
    (default) is recommended.}

\item{...}{Additional arguments. When already built dataset is used the User 
can define it with code{df}.}
}
\value{
A list containing the following components:
\itemize{
    \item{code{measure}} A string with the name of the measure column used.
    \item{code{method}} A string with the name of the method used.
    \item{code{baseline}} A string with the name of the baseline algorithm.
    \item{code{posteriror_probabilities}} A data frame with one row for every 
        algorithm that is compared to the baseline. The columns show the 
        posterior probabilities and whether significance appears.
    \item{code{repls}} Number of the considered replications.
}
}
\description{
This function implements a sequential approach of the Bayesian version of
    the signed rank test to compare the performance of machine learning 
    algorithms to one another. Sample size is not fixed in advance, data are 
    evaluated as they are collected. Further sampling is stopped in 
    accordance with a pre-defined stopping rule as soon as significant 
    results are obtained.
}
\details{
The basis for this test has first been implemented in rNPBST. For testing
    over multiple datasets, don´t specify the problem set argument in the 
    function. Note that if no measure column is defined per default the first
    column defined as measure_* in the data frame is used.
}
\examples{
    
    results <- seq_b_signed_rank_test(df = test_benchmark_small, 
    baseline = 'algo_1', max_repls = 10)
}
\references{
\url{https://github.com/JacintoCC/rNPBST}
}
