% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{test_benchmark_small}
\alias{test_benchmark_small}
\title{Benchmark Data}
\format{A data frame with 2500 rows and 5 variables:
\describe{
  \item{problem}{Problem set, where the data is tested on}
  \item{algorithm}{ML algorithms}
  \item{replications}{Number of replication, generated through sub-sampling}
  \item{measure_col}{Performance of algorithm measured with accuracy}
}}
\source{
Time series classification and regression with mlR.
}
\description{
Benchmark Data for multiple algorithms on a number of different data sets. 
(`test_benchmark_small`)
}
\references{
[seqbtests::test_benchmark_small]
}
