% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_tests.R
\name{b_corr_t_test}
\alias{b_corr_t_test}
\title{Bayesian correlated t test}
\usage{
b_corr_t_test(df, problem, baseline, algorithm = NULL, measure = NULL,
  compare = NULL, rho = 0.1, rope = c(-0.01, 0.01), prob = 0.95)
}
\arguments{
\item{df}{(`list`)\cr Data frame containing the performane measure.}

\item{problem}{(`character`)\cr Problem set used to evaluate the algorithms 
performance. Value in 'problem' column.}

\item{baseline}{(`character`)\cr First algorithm. Value in 'algorithm'
column.}

\item{algorithm}{(`character`)\cr Second algorithm. Value in 'algorithm' 
column. If not defined, the baseline is tested against all algorithms 
in the data frame.}

\item{measure}{(`character`)\cr Name of the 'measure' column. If not 
defined, the first 'measure' column in the data frame is used.}

\item{compare}{(`character`)\cr Defines if one algorithm needs to perform 
better ({\code{better}}) for decisions based on the posterior 
distribution or whether it is sufficient to perform not worse 
({\code{equal}}).}

\item{rho}{(`double`)\cr Correlation factor. Default is 0.1.}

\item{rope}{(`double`)\cr Region of practical equivalence. Default is 
c(-0.01, 0.01).}

\item{prob}{(`double`)\cr Threshold probability that decision rely on. 
Default is 0.95.}
}
\value{
(`list`)\cr A list containing the following components:
\itemize{
    \item{\code{measure}} (`character`)\cr A string with the name of the 
        measure column. 
    \item{\code{method}} (`character`)\cr A string with the name of the 
        method. 
    \item{\code{baseline}} (`character`)\cr A string with the name of the 
        first algorithm. Value in 'algorithm' column. 
    \item{\code{data_frame}} (`list`)\cr  A list containing the following 
        components:
    \item{\code{algorithm}} (`character`)\cr Second algorithm. Value in 
        'algorithm' column. If not defined, the baseline is tested against 
        all algorithms in the data frame. 
    \item{\code{left}} (`double`)\cr Left probability. 
    \item{\code{rope}} (`double`)\cr Rope probability. 
    \item{\code{right}} (`double`)\cr Right probability. 
    \item{\code{probabilities}} (`character`)\cr Decisions based on posterior 
        probabilities and threshold probability. 
}
}
\description{
This function implements the Bayesian version of the correlated t-test. 
    The performance of one baseline algorithm on one data set is compared to 
    either one or multiple algorithms.
}
\details{
The test has first been implemented in scmamp.
    If rho equals 0 this converts the test in the equivalent of the 
    standard t test.
}
\examples{
results <- b_corr_t_test(df= test_benchmark_small, problem = "problem_a", 
                         baseline = "algo_1", algorithm = "algo_2")
}
\references{
\url{https://github.com/b0rxa/scmamp}
}
