% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_tests.R
\name{b_corr_t_test}
\alias{b_corr_t_test}
\title{Bayesian correlated t test}
\usage{
b_corr_t_test(df, problem, baseline, algorithm = NULL, measure = NULL,
  compare = NULL, rho = 0.1, rope = c(-0.01, 0.01), prob = 0.95)
}
\arguments{
\item{df}{Input data frame.}

\item{problem}{Problem set on which the test should be performed.}

\item{baseline}{First algorithm.}

\item{algorithm}{Algorithm to be compared. If no algorithm is defined, the 
baseline is compared to every algorithm in the data frame.}

\item{measure}{Measure column.}

\item{compare}{Defines whether the baseline should be tested for either 
being better ('better') or being just as good ('equal') as the other 
algorithms. If not defined, the default is to test for 'better'.}

\item{rho}{Correlation factor.}

\item{rope}{Region of practical equivalence.}

\item{prob}{Probability, which the decision that the Baseline is better than 
the algorithm is based on. The default is 0.95.}
}
\value{
A list containing the following components:
\itemize{
    \item{code{measure}} A string with the name of the measure column used.
    \item{code{method}} A string with the name of the method used.
    \item{code{baseline}} A string with the name of the baseline algorithm.
    \item{code{posteriror_probabilities}} A data frame with one row for every 
        algorithm that is compared to the baseline. The columns show the 
        posterior probabilities and whether significance appears.
}
}
\description{
This function implements the Bayesian version of the correlated t-test. 
    The performance of one baseline algorithm on one data set is compared to 
    either one or multiple algorithms.
}
\details{
The test has first been implemented in scmamp. 
    Note that if no measure column is defined per default the first column 
    defined as measure_* in the data frame is used. The default of rho is 
    0.1. If rho equals 0 this converts the test in the equivalent of the 
    standard t test.
}
\examples{
results <- b_corr_t_test(df= test_benchmark_small, problem = "problem_a", 
                         baseline = "algo_1", algorithm = "algo_2")
}
\references{
\url{https://github.com/b0rxa/scmamp}
}
