% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_tests.R
\name{b_sign_test}
\alias{b_sign_test}
\title{Bayesian Sign test}
\usage{
b_sign_test(df, problem, baseline, algorithm = NULL, measure = NULL,
  compare = NULL, prob = 0.95, s = 1, z_0 = 0, weights = c(s/2,
  rep(1, length(x))), mc_samples = 1e+05, rope = c(-0.01, 0.01))
}
\arguments{
\item{df}{(`list`)\cr Data frame containing the performane measure.}

\item{problem}{(`character`)\cr Problem set used to evaluate the algorithms 
performance. Value in 'problem' column.}

\item{baseline}{(`character`)\cr First algorithm. Value in 'algorithm'
column.}

\item{algorithm}{(`character`)\cr Second algorithm. Value in 'algorithm' 
column. If not defined, the baseline is tested against all algorithms 
in the data frame.}

\item{measure}{(`character`)\cr Name of the 'measure' column. If not 
defined, the first 'measure' column in the data frame is used.}

\item{compare}{(`character`)\cr Defines if one algorithm needs to perform 
better ({\code{better}}) for decisions based on the posterior 
distribution or whether it is sufficient to perform not worse 
({\code{equal}}).}

\item{prob}{(`double`)\cr Threshold probability that decision rely on. 
Default is 0.95.}

\item{s}{(`double`)\cr Scale parameter of the prior Dirichlet Process. 
Default is 0.5}

\item{z_0}{(`double`)\cr Position of the pseudo-observation associated to 
the prior Dirichlet Process. Default is 0.}

\item{weights}{(`any`)\cr A prior weights.}

\item{mc_samples}{(`double`)\cr Number of samples used to estimate the 
posterior probability distribution.}

\item{rope}{(`double`)\cr Region of practical equivalence. Default is 
c(-0.01, 0.01).}
}
\value{
(`list`)\cr A list containing the following components:
\itemize{
    \item{\code{measure}} (`character`)\cr A string with the name of the 
        measure column. 
    \item{\code{method}} (`character`)\cr A string with the name of the 
        method. 
    \item{\code{baseline}} (`character`)\cr A string with the name of the 
        first algorithm. Value in 'algorithm' column. 
    \item{\code{data_frame}} (`list`)\cr  A list containing the following 
        components:
    \item{\code{algorithm}} (`character`)\cr Second algorithm. Value in 
        'algorithm' column. If not defined, the baseline is tested against 
        all algorithms in the data frame. 
    \item{\code{left}} (`double`)\cr Left probability. 
    \item{\code{rope}} (`double`)\cr Rope probability. 
    \item{\code{right}} (`double`)\cr Right probability. 
    \item{\code{probabilities}} (`character`)\cr Decisions based on posterior 
        probabilities and threshold probability. 
}
}
\description{
This function implements the Bayesian version of the sign test. The 
    performance of one baseline algorithm on one or multiple data sets is 
    compared to either one or multiple algorithms.
}
\details{
The test has first been implemented in rNPBST.
}
\examples{
    results <- b_sign_test(df= test_benchmark_small, 
    problem = "problem_a", baseline = "algo_1", algorithm = "algo_2")
}
\references{
\url{https://github.com/JacintoCC/rNPBST}
}
