% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_tests.R
\name{b_sign_test}
\alias{b_sign_test}
\title{Bayesian Sign test}
\usage{
b_sign_test(df, problemset, baseline, algorithm = NULL, measure = NULL,
  compare = NULL, s = 1, z_0 = 0, weights = c(s/2, rep(1,
  length(x))), mc_samples = 1e+05, rope = c(-0.01, 0.01))
}
\arguments{
\item{df}{Input data frame.}

\item{problemset}{Problem set on which the test should be performed.}

\item{baseline}{First algorithm.}

\item{algorithm}{Second algorithm. If not defined, every algorithm will be 
tested against baseline.}

\item{measure}{Measure column.}

\item{s}{Prior pseudo-observation probability.}

\item{z_0}{Prior pseudo-observation.}

\item{weights}{A-priori weights.}

\item{mc_samples}{Number of samples of the distribution.}

\item{rope}{Region of practical equivalence.}
}
\description{
This function implements the Bayesian version of the sign test. The 
    performance of one baseline algorithm on one or multiple data sets is 
    compared to either one or multiple algorithms.
}
\details{
The test has first been implemented in rNPBST. For testing over multiple 
    datasets, don´t specify the problem set argument in the function. Note 
    that if no measure column is defined per default the first column 
    defined as measure_* in the data frame is used.
}
\references{
\url{https://github.com/JacintoCC/rNPBST}
}
