% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_structure.R
\name{check_column_names}
\alias{check_column_names}
\title{Check Columns of Data Frame}
\usage{
check_column_names(df)
}
\arguments{
\item{df}{input data frame}
}
\value{
TRUE if test was successful
}
\description{
Check if data frame only contains legit columns
  and whether the compulsory columns exist.
}
