% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_structure.R
\name{check_names}
\alias{check_names}
\title{Check column names}
\usage{
check_names(df, problemset = NULL, baseline = NULL, learner_a = NULL,
  learner_b = NULL, measure = NULL, parameter_algorithm = NULL)
}
\description{
Check if column names that were passed by User
  are contained in the data frame.
}
