% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_structure.R
\name{check_names}
\alias{check_names}
\title{Check column names}
\usage{
check_names(df, problem = NULL, baseline = NULL, algorithm = NULL,
  measure = NULL)
}
\arguments{
\item{df}{(`list`)\cr Data frame containing the performane measure.}

\item{problem}{(`character`)\cr Problem set used to evaluate the algorithms 
performance. Value in 'problem' column.}

\item{baseline}{(`character`)\cr First algorithm. Value in 'algorithm'
column.}

\item{algorithm}{(`character`)\cr Second algorithm. Value in 'algorithm' 
column. If not defined, the baseline is tested against all algorithms 
in the data frame.}

\item{measure}{(`character`)\cr Name of the 'measure' column. If not 
defined, the first 'measure' column in the data frame is used.}
}
\value{
(`logical`)\cr TRUE if test was successful.
}
\description{
Check if column names that were passed by User are contained 
    in the data frame.
}
\examples{
    check_names(test_benchmark_small)
}
