% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_structure.R
\name{check_structure}
\alias{check_structure}
\title{Check Structure}
\usage{
check_structure(df)
}
\arguments{
\item{df}{input data frame}
}
\value{
TRUE if test was successful
}
\description{
Check if the structure of the data frame satisfies the requriements.
}
