% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequentist_tests.R
\name{corr_t_test}
\alias{corr_t_test}
\title{Correlated t Test}
\usage{
corr_t_test(df, problemset, learner_a, learner_b, measure = NULL,
  rho = 0.01, alternative = "two.sided")
}
\arguments{
\item{df}{Input data frame.}

\item{problemset}{Problemset on which the test should be performed.}

\item{learner_a}{First algorithm.}

\item{learner_b}{Second algorithm.}

\item{measure}{Measure column.}

\item{rho}{Correlation factor. (for the case of cross validated results, the 
heuristic to set the correlation is size of test set divided by total size 
of the dataset)}

\item{alternative}{a character string specifying the alternative hypothesis, 
must be one of "two.sided" (default), "greater" or "less". 
You can specify just the initial letter.}
}
\value{
A list containing the following components: 
\item{code{measure}}{a string with the name of the measure column used}
\item{code{method}}{a string with the name of the method used}
\item{code{statistic}}{the value of the statistic used in the test}
\item{code{p_value}}{the p-value for the test}
}
\description{
This function implements the t-Test for paired samples.
}
\details{
The test has first been implemented in scmamp.
Note that the default value for measure is the first measure column in the 
data frame.
}
\references{
\url{https://github.com/b0rxa/scmamp}
}
