% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequentist_tests.R
\name{corr_t_test}
\alias{corr_t_test}
\title{Correlated t test}
\usage{
corr_t_test(df, problem, baseline, algorithm = NULL, measure = NULL,
  rho = 0.01)
}
\arguments{
\item{df}{(`list`)\cr Data frame containing the performane measure.}

\item{problem}{(`character`)\cr Problem set used to evaluate the algorithms 
performance. Value in 'problem' column.}

\item{baseline}{(`character`)\cr First algorithm. Value in 'algorithm'
column.}

\item{algorithm}{(`character`)\cr Second algorithm. Value in 'algorithm' 
column. If not defined, the baseline is tested against all algorithms 
in the data frame.}

\item{measure}{(`character`)\cr Name of the 'measure' column. If not 
defined, the first 'measure' column in the data frame is used.}

\item{rho}{(`double`)\cr Correlation factor. Default is 0.1. (For the case of 
cross validated results, the heuristic to set the correlation is size of 
test set divided by total size of the dataset.)}
}
\value{
A list containing the following components: 
\itemize{
    \item{\code{measure}} (`character`)\cr A string with the name of the 
        measure column. 
    \item{\code{method}} (`character`)\cr A string with the name of the 
        method. 
    \item{\code{statistic}} (`double`)\cr The value of the statistic used in 
        the test.
    \item{\code{p_value}} (`double`)\cr The p-value for the test.
}
}
\description{
This function implements a two-sided t test for paired samples.
}
\details{
The test has first been implemented in scmamp.
}
\examples{
    results <- corr_t_test(df= test_benchmark_small, 
    problem = "problem_a", baseline = "algo_1", algorithm = "algo_2")
}
\references{
\url{https://github.com/b0rxa/scmamp}
}
