% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_transformation}
\alias{data_transformation}
\title{Data transformation}
\usage{
data_transformation(df, algo, measure)
}
\arguments{
\item{df}{Input data frame.}

\item{algo}{Name of the algorithm that shall be compared.}

\item{measure}{Measure column.}
}
\value{
Matrix.
}
\description{
Extract a data matrix out of the original dataset. The matrix 
    contains all observations for each replication in each problem set. 
    The replications are stored in the rows, while the columns are according  
    to the problem sets.
}
