% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_hoc_tests.R
\name{friedman_post}
\alias{friedman_post}
\title{Friedman's post hoc test}
\usage{
friedman_post(df, measure = NULL, control = NULL)
}
\arguments{
\item{df}{Input data frame.}

\item{measure}{Measure column.}

\item{control}{The name of the control algorithm. If this parameter is not 
provided, all algorithms are compared against each other.}
}
\description{
This function implements a Friedman post hoc test. It computes 
the raw p-values for the post hoc based on Friedman's test.
}
\details{
The test has first been implemented in scmamp. 
    Note that if no measure column is defined per default the first column 
    defined as measure_* in the data frame is used.
}
\references{
\url{https://github.com/b0rxa/scmamp}
}
