% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequentist_tests.R
\name{friedman_test}
\alias{friedman_test}
\title{Friedman's test}
\usage{
friedman_test(df, measure = NULL)
}
\arguments{
\item{df}{Input data frame.}

\item{measure}{Measure column.}
}
\value{
A list containing the following components: 
\item{code{measure}}{a string with the name of the measure column used}
\item{code{method}}{a string with the name of the method used}
\item{code{statistic}}{the value of the statistic used in the test}
\item{code{p_value}}{the p-value for the test}
}
\description{
This function implements the Friedman's test for multiple 
comparisons.A non-parametric statistical test zo detect differences in 
in algorithms performances over multiple datasets.
}
\details{
The test has first been implemented in scmamp.
Note that the default value for measure is the first measure column in the 
data frame.
}
\references{
\url{https://github.com/b0rxa/scmamp}
}
