% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequentist_tests.R
\name{friedman_test}
\alias{friedman_test}
\title{Friedman's test}
\usage{
friedman_test(df, measure = NULL)
}
\arguments{
\item{df}{(`list`)\cr Data frame containing the performane measure.}

\item{measure}{(`character`)\cr Name of the 'measure' column. If not 
defined, the first 'measure' column in the data frame is used.}
}
\value{
A list containing the following components: 
\itemize{
    \item{\code{measure}} (`character`)\cr A string with the name of the 
        measure column. 
    \item{\code{method}} (`character`)\cr A string with the name of the 
        method. 
    \item{\code{statistic}} (`double`)\cr The value of the statistic used in 
        the test.
    \item{\code{p_value}} (`double`)\cr The p-value for the test.
}
}
\description{
This function implements the Friedman's test for multiple 
    comparisons. A non-parametric statistical test to detect differences in 
    in algorithms performances over multiple datasets.
}
\details{
The test has first been implemented in scmamp.
}
\examples{
    results <- friedman_test(test_benchmark) 
}
\references{
\url{https://github.com/b0rxa/scmamp}
}
