% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data_frame}
\alias{get_data_frame}
\title{Get test results (for data frame)}
\usage{
get_data_frame(k, posterior)
}
\arguments{
\item{k}{Algorithm tested against the baseline.}

\item{posterior}{Call for the posterior probabilities of the bayesian tests.}
}
\value{
List.
}
\description{
This function collects the part of the results shown in the data 
    frame.
}
