% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data_frame_htest}
\alias{get_data_frame_htest}
\title{Get test results (for frequentist tests)}
\usage{
get_data_frame_htest(k, p_value, test, statistic)
}
\arguments{
\item{k}{(`character`)\cr Algorithm tested against the baseline.}

\item{p_value}{(`double`)\cr The p-value for the test.}

\item{test}{(`character`)\cr Hypothesis test.}

\item{statistic}{(`double`)\cr Value of the statistic used in the test.}
}
\value{
(`list`)\cr List.
}
\description{
This function collects the part of the results shown in the data 
    frame.
}
