% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data_frame_htest_small}
\alias{get_data_frame_htest_small}
\title{Get test results (for frequentist tests)}
\usage{
get_data_frame_htest_small(p_value, test, statistic)
}
\value{
List.
}
\description{
This function collects the part of the results shown in the data 
    frame. Not depending on algorithms.
}
