% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data_frame_seq}
\alias{get_data_frame_seq}
\title{Get test results (for sequential tests)}
\usage{
get_data_frame_seq(k, posterior, repls)
}
\arguments{
\item{k}{(`character`)\cr Algorithm tested against the baseline.}

\item{posterior}{(`character`)\cr Call for the posterior probabilities of the 
Bayesian tests.}

\item{repls}{(`double`)\cr Number of replications used until a decision is made.}
}
\value{
(`list`)\cr List.
}
\description{
This function collects the part of the results shown in the data 
    frame.
}
