% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_main_columns}
\alias{get_main_columns}
\title{Main Columns}
\usage{
get_main_columns(df)
}
\arguments{
\item{df}{input data frame}
}
\value{
A vector containing all main column names.
}
\description{
Get the list of names of the main columns within the data frame, 
  namely "problem", "algorithm" and "replications".
}
