% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_main_columns_count}
\alias{get_main_columns_count}
\title{Count Main Columns}
\usage{
get_main_columns_count(df)
}
\arguments{
\item{df}{input data frame}
}
\value{
A numeric vector displaying the number of main columns.
}
\description{
Count the number of main columns in the data frame.
}
