% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_measure_columns}
\alias{get_measure_columns}
\title{Measure columns}
\usage{
get_measure_columns(df)
}
\arguments{
\item{df}{Input data frame.}
}
\value{
A vector containing all measure columns.
}
\description{
Get the list of names for all columns containing performance 
    measures of the algorithms within the data frame.
}
