% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_measure_columns_count}
\alias{get_measure_columns_count}
\title{Count measure columns}
\usage{
get_measure_columns_count(df)
}
\arguments{
\item{df}{(`list`)\cr Data frame containing the performane measure.}
}
\value{
(`list`)\cr A numeric vector displaying the number of columns  
    containing performance measures.
}
\description{
Get the number parameter columns in the data frame.
}
