% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_parameter_columns}
\alias{get_parameter_columns}
\title{Parameter columns}
\usage{
get_parameter_columns(df)
}
\arguments{
\item{df}{Input data frame.}
}
\value{
A vector containing all parameter columns.
}
\description{
Get the list of names for all parameter columns within the 
    data frame.
}
