% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_parameter_columns_count}
\alias{get_parameter_columns_count}
\title{Count parameter columns}
\usage{
get_parameter_columns_count(df)
}
\arguments{
\item{df}{(`list`)\cr Data frame containing the performane measure.}
}
\value{
(`list`)\cr A numeric vector displaying the number of parameter 
    columns.
}
\description{
Get the number parameter columns in the data frame.
}
