% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_probabilities}
\alias{get_probabilities}
\title{Get probabilities}
\usage{
get_probabilities(result, compare, prob)
}
\arguments{
\item{result}{Data frame with all the results within in the tests.}

\item{compare}{Defines whether the baseline should be tested for either 
being better ('better') or being just as good ('equal') as the other 
algorithms.}

\item{prob}{Probability, which the decision that the Baseline is better than 
the algorithm is based on.}
}
\value{
Column of result data frame.
}
\description{
This function defines the probabilities resulting from the
    Bayesian tests for each algorithm that is tested against the baseline.
}
