% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_replications}
\alias{get_replications}
\title{Build Replications}
\usage{
get_replications(i, df)
}
\arguments{
\item{i}{number of replications used}

\item{df}{input data frame}
}
\value{
Dataframe containing defined number of replications.
}
\description{
When not defined by user, it calls the replications in a complete dataset.  
Can be used to build replications during the testing.
}
