% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_results}
\alias{get_results}
\title{Get test results (Bayesian tests)}
\usage{
get_results(baseline, method, measure, data = NULL, extra = NULL)
}
\arguments{
\item{baseline}{(`character`)\cr First algorithm. Value in 'algorithm'
column.}

\item{method}{(`character`)\cr Bayesian test.}

\item{measure}{(`character`)\cr Name of the 'measure' column. If not 
defined, the first 'measure' column in the data frame is used.}

\item{data}{(`list`)\cr Data frame containing the posterior probabilities.}

\item{extra}{(`any`)\cr Any extra arguments needed to build the plots.}
}
\value{
(`list`)\cr List.
}
\description{
This function collects the results of the Bayesian tests.
}
