% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_results_htest}
\alias{get_results_htest}
\title{Get test results (Frequentist tests)}
\usage{
get_results_htest(baseline = NULL, method, measure, data = NULL,
  matrix = NULL)
}
\arguments{
\item{baseline}{Baseline algorithms that every algorithm can be compared to.}

\item{method}{Frequentist test.}

\item{measure}{Measure column.}

\item{data}{Data frame containing the posterior probabilities.}

\item{matrix}{A matrix with all the pair wise differences of 
average rankings.}
}
\value{
List.
}
\description{
This function collects the results of the Frequentist tests.
}
