% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_threshold}
\alias{get_threshold}
\title{Get threshold}
\usage{
get_threshold(posterior, compare)
}
\arguments{
\item{posterior}{Call for the posterior probabilities of the bayesian tests.}

\item{compare}{Defines whether the baseline should be tested for either 
being better}
}
\description{
This function defines the threshold for making any decisions 
    depending on the calculated posterior probabilities.
}
