% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_function.R
\name{na_check}
\alias{na_check}
\title{NA Check}
\usage{
na_check(df, measure, check_var)
}
\arguments{
\item{df}{input data frame}

\item{measure}{Measure column}

\item{check_var}{Column in data frame used to check for NAs}
}
\value{
List of Cases, NAs and the NA ratio according to the check_var values.
}
\description{
Check if the measure column is complete. For values of one of the other 
columns this function shows the Ratio of NAs existing in the measure column. 
If there are any NAs the User can decide to drop all observations for that 
specific value, since the data frame needs to be complete for testing.
}
