% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_function.R
\name{na_drop}
\alias{na_drop}
\title{Drop NAs by groups}
\usage{
na_drop(df, check_var, measure)
}
\arguments{
\item{df}{Input data frame.}

\item{check_var}{Column in data frame used to check for NAs. Either 
"problemsets" or "algorithm".}

\item{measure}{Measure column.}
}
\value{
New data frame without NAs.
}
\description{
Drop group of rows that contain any NA depending on values of 
check_var.
}
