% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_hoc_tests.R
\name{nemenyi_test}
\alias{nemenyi_test}
\title{Nemenyi post hoc test}
\usage{
nemenyi_test(df, measure = NULL, alpha = 0.05)
}
\arguments{
\item{df}{Input data frame.}

\item{measure}{Measure column.}

\item{alpha}{Significance level.}
}
\value{
A list containing the following components: 
\itemize{
    \item{code{measure}} A string with the name of the measure column used.
    \item{code{method}} A string with the name of the method used.
    \item{code{statistic}} The value of the statistic used in the test.
    \item{code{p_value}} The p-value for the test.
    \item{code{diff_matrix}} A matrix with all the pair wise differences of 
    average rankings.
}
}
\description{
This function implements the Nemenyi post hoc test.
}
\details{
The test has first been implemented in scmamp. 
    Note that if no measure column is defined per default the first column 
    defined as measure_* in the data frame is used.
}
\examples{
    nemenyi_test(test_benchmark)
}
\references{
\url{https://github.com/b0rxa/scmamp}
}
