% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_hoc_tests.R
\name{nemenyi_test}
\alias{nemenyi_test}
\title{Nemenyi post hoc test}
\usage{
nemenyi_test(df, measure = NULL, alpha = 0.05)
}
\arguments{
\item{df}{Input data frame.}

\item{measure}{Measure column.}

\item{alpha}{Significance level.}
}
\value{
A list containing the following components: 
\item{code{measure}}{A string with the name of the measure column used.}
\item{code{method}}{A string with the name of the method used.}
\item{code{statistic}}{The value of the statistic used in the test.}
\item{code{p_value}}{The p-value for the test.}
\item{code{diff_matrix}}{A matrix with all the pair wise differences of 
average rankings.}
\item{code{significance}}{A matrix with all the pair wise comparisons showing 
if there are significant differences among the algorithms}
}
\description{
This function implements the Nemenyi post hoc test.
}
\details{
The test has first been implemented in scmamp. Note that the default
value for measure is the first measure column in the data frame.
}
\references{
\url{https://github.com/b0rxa/scmamp}
}
