% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_structure.R
\name{paste_algo_pars}
\alias{paste_algo_pars}
\title{Paste algorithm and parameter}
\usage{
paste_algo_pars(df, parameter_algo = NULL)
}
\arguments{
\item{df}{(`list`)\cr Data frame containing the performane measure.}

\item{parameter_algo}{df (`character`)\cr Algorithm parameter in data 
frame.}
}
\value{
df (`list`)\cr New data frame, where the algorithm and 
    parameter_algo column, as well as the value names are combined.
}
\description{
If there is a parameter additional to the algorithm, both can 
    be combined, when using this function.
}
