% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{Boxplot}
\usage{
plot_boxplot(df, measure = NULL)
}
\arguments{
\item{df}{Input data frame.}

\item{measure}{Measure column.}
}
\description{
This function plots the performance of every algorithm in each 
    data frame in boxplots.
}
\details{
The test has first been implemented in scmamp.
}
\examples{
    plot_boxplot(df = test_benchmark_small)
}
\references{
\url{https://github.com/b0rxa/scmamp}
}
