% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{Boxplot}
\usage{
plot_boxplot(df, measure = NULL)
}
\arguments{
\item{df}{Input data frame.}

\item{measure}{Measure column.}
}
\value{
An object of class \linkS4class{ggplot} with the plot
}
\description{
This function plots the performance of all algorithms in the 
data frame as boxplots
}
\details{
The test has first been implemented in scmamp.
}
\references{
\url{https://github.com/b0rxa/scmamp}
}
