% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_cd}
\alias{plot_cd}
\title{Critical differences plot}
\usage{
plot_cd(df, measure = NULL, alpha = 0.05, cex = 0.75, ...)
}
\arguments{
\item{df}{Input data frame.}

\item{measure}{Measure column.}

\item{alpha}{Significance level to get the critical difference.}

\item{cex}{Numeric value to control the size of the font.}
}
\description{
This function plots the critical differences plot.
}
\details{
The test has first been implemented in scmamp.
Note that the default value for measure is the first measure column in the 
data frame. By default, the alpha value is 0.05 and the default for cex is 
0.75.
}
\references{
\url{https://github.com/b0rxa/scmamp}
}
