% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_posterior}
\alias{plot_posterior}
\title{Posterior densities plot}
\usage{
plot_posterior(df, problemset, learner_a, learner_b, measure = NULL,
  rho = 0.1, rope = c(-0.01, 0.01), parameter = 1, points = 1000,
  plot_rope = TRUE, plot_samples = TRUE, alpha = NULL)
}
\arguments{
\item{df}{Input data frame.}

\item{problemset}{Problemset on which the test should be performed.}

\item{learner_a}{First algorithm.}

\item{learner_b}{Second algorithm.}

\item{measure}{Measure column.}

\item{rho}{Correlation factor.}

\item{rope}{Region of practical equivalence.}

\item{parameter}{Either a string or a number indicating, in case the 
posterior is approximated, the parameter to be plotted (i.e., the name or the 
index of a column in the sample matrix)}

\item{points}{Number of points used to plot the functions}

\item{plot_rope}{A logical value indicating whether the rope has to be 
plotted or not. Note that not for all
parameter the rope makes sense}

\item{plot_samples}{A logical value. If true, the samples are plotted (only 
when the posterior is approximate)}

\item{alpha}{Numeric value for the transparency of the points, only 
applicable if \code{plot.samples} is true}

\item{...}{Additional parameters to the Rgraphviz function. This is mainly 
to change the layout of the graph}
}
\value{
An object of class \linkS4class{ggplot} with the plot
}
\description{
This function plots, univariately, the posterior densities 
obtained in Bayesian analyses.
}
\details{
The test has first been implemented in scmamp.
}
\references{
\url{https://github.com/b0rxa/scmamp}
}
