% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\name{plot_triangles}
\alias{plot_triangles}
\title{Plot Posterior}
\usage{
plot_triangles(x, num.points = nrow(x$sample), ...)
}
\arguments{
\item{x}{Results of a Bayesian test.}

\item{num.points}{Number of points to be shown in the plot.}
}
\description{
Plot of the projection of 3-simplex points of a posterior 
    Dirichlet distribution into a 2D triangle. The function is used from the 
    package rNPBST. Since it is a numeric function, the code was copied for
    further use in this package.
}
\references{
\url{https://github.com/JacintoCC/rNPBST}
}
