% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_sequential_tests.R
\name{seq_b_corr_t_test}
\alias{seq_b_corr_t_test}
\title{Sequential Bayesian correlated t-test}
\usage{
seq_b_corr_t_test(
  problem,
  baseline,
  algorithm = NULL,
  measure = NULL,
  compare = NULL,
  rho = 0.1,
  rope = c(-0.01, 0.01),
  max_repls = 20,
  prob = 0.95,
  min_repls = 5,
  ...
)
}
\arguments{
\item{problem}{(`character`)\cr Problem set used to evaluate the algorithms 
performance. Value in 'problem' column.}

\item{baseline}{(`character`)\cr First algorithm. Value in 'algorithm'
column.}

\item{algorithm}{(`character`)\cr Second algorithm. Value in 'algorithm' 
column. If not defined, the baseline is tested against all algorithms 
in the data frame.}

\item{measure}{(`character`)\cr Name of the 'measure' column. If not 
defined, the first 'measure' column in the data frame is used.}

\item{compare}{(`character`)\cr Defines if one algorithm needs to perform 
better ({\code{better}}) for decisions based on the posterior 
distribution or whether it is sufficient to perform not worse 
({\code{equal}}).}

\item{rho}{(`double`)\cr Correlation factor. Default is 0.1.}

\item{rope}{(`double`)\cr Region of practical equivalence. Default is 
c(-0.01, 0.01).}

\item{max_repls}{(`double`)\cr Maximum number of replications. If a complete 
data frame is passed ({\code{max_repls}}) should correspond to the 
maximum number of replications that are built. Default is 20.}

\item{prob}{(`double`)\cr Threshold probability that decision rely on. 
Default is 0.95.}

\item{min_repls}{(`double`)\cr Minimum number of replications that is 
used/generated before an optional stopping rule is activated. Default is 
5.}

\item{...}{(any)\cr Additional arguments for ({\code{get_replication}}). 
To pass a complete data frame, set ({\code{df}}) (`character`).}
}
\value{
(`list`)\cr A list containing the following components:
\itemize{
    \item{\code{measure}} (`character`)\cr A string with the name of the 
        measure column. 
    \item{\code{method}} (`character`)\cr A string with the name of the 
        method. 
    \item{\code{baseline}} (`character`)\cr A string with the name of the 
        first algorithm. Value in 'algorithm' column. 
    \item{\code{data_frame}} (`list`)\cr  A list containing the following 
        components:
    \item{\code{algorithm}} (`character`)\cr Second algorithm. Value in 
        'algorithm' column. If not defined, the baseline is tested against 
        all algorithms in the data frame. 
    \item{\code{left}} (`double`)\cr Left probability. 
    \item{\code{rope}} (`double`)\cr Rope probability. 
    \item{\code{right}} (`double`)\cr Right probability. 
    \item{\code{repls}} (`double`)\cr Number of evaluated replications. 
    \item{\code{probabilities}} (`character`)\cr Decisions based on posterior 
        probabilities and threshold probability. 
}
}
\description{
This function implements a sequential approach of the Bayesian correlated 
    t-test to compare the performance of machine learning algorithms to one 
    another. Sample size is not fixed in advance, data are evaluated as they 
    are collected. Further sampling is stopped in accordance with a 
    pre-defined stopping rule.
}
\details{
The basis for this test has first been implemented in scmamp.
}
\examples{
    results <- seq_b_corr_t_test(df = test_benchmark_small, rho=0.1,
    problem = "problem_a", baseline = "algo_1", 
    compare = "equal", max_repls = 10)
}
\references{
\url{https://github.com/b0rxa/scmamp}
}
