% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_sequential_tests.R
\name{seq_b_corr_t_test}
\alias{seq_b_corr_t_test}
\title{Sequential bayesian correlated t test}
\usage{
seq_b_corr_t_test(problemset, baseline = NULL, learner_b = NULL,
  measure = NULL, compare = NULL, rho = 0.1, rope = c(-0.01, 0.01),
  max_repls = 20, ...)
}
\arguments{
\item{problemset}{Problemset on which the test should be performed.}

\item{baseline}{First algorithm.}

\item{learner_b}{Second algorithm. If not defined, every algorithm will be 
tested against baseline.}

\item{measure}{Measure column.}

\item{compare}{Defines whether the performances should be tested for either 
being better ('better') or being just as good ('equal'). If not defined, the 
default is to test for 'better'.}

\item{rho}{Correlation factor.}

\item{rope}{Region of practical equivalence.}

\item{max_repls}{maximum number of replications that should be build in 
get_replications, or maximum number of replications in data frame, if 
complete data frame is being used.}

\item{rope}{Region of practical equivalence.}
}
\value{
A data frame with one row for each considered algorithm that is 
tested against the baseline, containing the following components:
\item{code{measure}}{a string with the name of the measure column used}
\item{code{method}}{a string with the name of the method used}
\item{code{posteriror_probabilities}}{a vector with the left, rope and right 
probabilities}
\item{code{repls}}{number of the considered replications}
}
\description{
This function implements a sequential approach to compare the performance of 
machine learning algorithms to one another. Sample size is not fixed in 
adavnce, data are evaluated as they are collected. Further sampling is 
stopped in accordance with a pre-defined stopping rule as soon as significant 
results are obtained.
}
\details{
The basics of the test have first been implemented in scmamp. 
Note that the default value for measure is the first measure column in the 
data frame. The default of rho is 0.1. If rho equals 0 this converts the test 
in the equivalent of the standard t test.
}
