% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_sequential_tests.R
\name{seq_b_hierarchical_test}
\alias{seq_b_hierarchical_test}
\title{Sequential Bayesian hierarchical correlated t-test}
\usage{
seq_b_hierarchical_test(baseline, algorithm = NULL, measure = NULL,
  compare = NULL, rho = 0.1, max_repls = 20, rope = c(-0.01, 0.01),
  std.upper = 1000, d0.lower = NULL, d0.upper = NULL,
  alpha.lower = 0.5, alpha.upper = 5, beta.lower = 0.05,
  beta.upper = 0.15, nsim = 2000, nchains = 8, parallel = TRUE,
  stan.output.file = NULL, seed = as.numeric(Sys.time()), ...)
}
\arguments{
\item{baseline}{First algorithm.}

\item{algorithm}{Second algorithm. If not defined, every algorithm will be 
tested against baseline.}

\item{measure}{Measure column.}

\item{compare}{Defines whether the baseline should be tested for either 
being better ('better') or being just as good ('equal') as the other 
algorithms. If not defined, the default is to test for 'better'.}

\item{rho}{Correlation factor.}

\item{max_repls}{Maximum number of replications that should be build in 
get_replications. Or maximum number of replications in data frame, if 
a data frame that has already been built is being used.}

\item{rope}{Interval for the difference considered as 'irrelevant'.}

\item{std.upper}{Factor to set the upper bound for both sigma_i and sigma_0.}

\item{d0.lower}{Lower bound for the prior for mu_0. If not provided, 
the smallest observed difference is used.}

\item{d0.upper}{Upper bound for the prior for mu_0. If not provided, 
the biggest observed difference is used.}

\item{alpha.lower}{Lower bound for the (uniform) prior for the alpha 
hyperparameter. Default value set at 0.5.}

\item{alpha.upper}{Upper bound for the (uniform) prior for the alpha 
hyperparameter. Default value set at 0.5.}

\item{beta.lower}{Lower bound for the (uniform) prior for the beta 
hyperparameter. Default value set at 0.5.}

\item{nsim}{Number of samples (per chain) used to estimate the posterior 
distribution. Note that, by default, half the simulations are used for 
the burn-in.}

\item{parallel}{Logical value. If \code{true}, Stan code is executed in 
parallel.}

\item{stan.output.file}{String containing the base name for the output files 
produced by Stan. If \code{NULL}, no files are stored.}

\item{seed}{Optional parameter used to fix the random seed}

\item{...}{Additional arguments. When already built dataset is used the User 
can define it with code{df}.}

\item{beta.lower}{Upper bound for the (uniform) prior for the beta 
hyperparameter. Default value set at 0.5.}

\item{z0}{Position of the pseudo-observation associated to the prior 
Dirichlet Process. The default value is set to 0 (inside the rope).}

\item{nchain}{Number of MC chains to be simulated. As half the simulations 
are used for the warm-up, the total number of simulations will 
be \code{nchain}*\code{nsim}/2.}

\item{rope}{Region of practical equivalence.}
}
\description{
This function implements a sequential approach of the Bayesian 
    hierarchical test to compare the performance of machine learning 
    algorithms to one another. Sample size is not fixed in advance, data are 
    evaluated as they are collected. Further sampling is stopped in 
    accordance with a pre-defined stopping rule as soon as significant 
    results are obtained.
}
\details{
The basis for this test has first been implemented in scmamp. 
    Note that if no measure column is defined per default the first column 
    defined as measure_* in the data frame is used. The default of rho is 
    0.1.
}
\references{
\url{https://github.com/b0rxa/scmamp}
}
