% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequentist_tests.R
\name{wilcoxon_signed_test}
\alias{wilcoxon_signed_test}
\title{Wilcoxon signed-rank test}
\usage{
wilcoxon_signed_test(df, problem, baseline, algorithm = NULL, measure = NULL)
}
\arguments{
\item{df}{(`list`)\cr Data frame containing the performane measure.}

\item{problem}{(`character`)\cr Problem set used to evaluate the algorithms 
performance. Value in 'problem' column.}

\item{baseline}{(`character`)\cr First algorithm. Value in 'algorithm'
column.}

\item{algorithm}{(`character`)\cr Second algorithm. Value in 'algorithm' 
column. If not defined, the baseline is tested against all algorithms 
in the data frame.}

\item{measure}{(`character`)\cr Name of the 'measure' column. If not 
defined, the first 'measure' column in the data frame is used.}
}
\value{
A list containing the following components: 
\itemize{
    \item{\code{measure}} (`character`)\cr A string with the name of the 
        measure column. 
    \item{\code{method}} (`character`)\cr A string with the name of the 
        method. 
    \item{\code{statistic}} (`double`)\cr The value of the statistic used in 
        the test.
    \item{\code{p_value}} (`double`)\cr The p-value for the test.
}
}
\description{
This function implements the paired Wilcoxon signed-rank test. A 
    non-parametric statistical hypothesis test to compare the means of two 
    paired samples.
}
\details{
The test has first been implemented in scmamp.
}
\examples{
    results <- wilcoxon_signed_test(df = test_benchmark, baseline = "algo_1",
    algorithm = "algo_2", problem = "problem_a")  
}
\references{
\url{https://github.com/b0rxa/scmamp}
}
