% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequentist_tests.R
\name{wilcoxon_signed_test}
\alias{wilcoxon_signed_test}
\title{Wilcoxon signed-rank test}
\usage{
wilcoxon_signed_test(df, problemset, baseline, algorithm, measure = NULL)
}
\arguments{
\item{df}{Input data frame.}

\item{problemset}{Problem set on which the test should be performed.}

\item{baseline}{First algorithm.}

\item{algorithm}{Second algorithm.}

\item{measure}{Measure column.}
}
\value{
A list containing the following components: 
\item{code{measure}}{A string with the name of the measure column used.}
\item{code{method}}{A string with the name of the method used.}
\item{code{statistic}}{The value of the statistic used in the test.}
\item{code{p_value}}{The p-value for the test.}
}
\description{
This function implements the paired Wilcoxon signed-rank test. A 
non-parametric statistical hypothesis test to compare the means of two 
paired samples.
}
\details{
The test has first been implemented in scmamp.
Note that the default value for measure is the first measure column in the 
data frame.
}
\references{
\url{https://github.com/b0rxa/scmamp}
}
